ml_project/requirements.txt
numpy
click
PyYAML
joblib
pandas
scikit-learn
marshmallow_dataclass